feat(storage): refactor storage handling and migrate article storage to a new module
